pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "SerialTerminal"
version = "0.1.0"
description = "TUI Serial terminal"
requires-python = ">=3.9"

dependencies = [
    "textual>=0.47.0",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    ""
]

[project.scripts]
serialterminal = "main:main"

[tool.setuptools]
packages = [
    "ui",
    "config",
    "models",
    "serial_comm",
    "utils",
]
